Implemented basic repeating functionality.
Summary:
Added the structure necessary for repeating alarms, mainly centered
around a new wrapper class, AlarmRepeats. Interface to this class
is documented for all functions.

Additionally, this commit includes some useful general functionality
for the Alarm class; in particular, functions to adjust the Date of
the Alarm and reset it to the next occurrence of a particular time.

Finally, the rudimentary GUI has been updated with checkboxes to
permit selection of repeat days.